backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
replicate==0.18.1
pydantic>=2.5.2
httpx==0.25.1
